backend/requirements.txt
fastapi
uvicorn
crewai
anthropic
ultralytics
opencv-python
pymongo
python-multipart
pydantic
python-dotenv
websockets
